Fix `clear pane` buttons not working after ajax reload
When using `searchPanes` and `table.ajax.reload()`, the "clear pane" buttons available on each pane if the `clear` option is set to `true`, stops working as the event listeners are only set the first time the pane is displayed. This fix makes sure to reset the flag when rebuilding the pane so that the event listeners are re-attached as expected. (one-liner fix at line [332](https://github.com/DataTables/SearchPanes/edit/master/src/searchPane.ts#diff-ea1d751cf1d988a555a4824991527f1eR332))

The other changes are not mandatory, it's just to make sure there's no event listeners that are added twice or more as this was causing memory leaks and performance degradation in previous versions.