Fix: Activating a search on a pane should reset the paging of the host table. Otherwise it was possible to end up when incorrect paging.
This reverts my change from 3139bf17d167c95dc3f24ad6aedd387804b7b823.
The more I think about it, the more I think that original commit and the
decision behind it was wrong. If the data has changed, then it should
redraw the table (the filtering might have changed as a result).